完成了消息模块的模型定义，api，实现controller

+++ mes-java/mes-apis/src/main/java/com/zeroone/star/project/j3/sysmanager/MessageManageApis.java
@@ -17,7 +17,7 @@ public interface MessageManageApis {
      *
      * @return 消息详情
      */
-    JsonVO<NewsVO> getNewsById(Integer id);
+    JsonVO<NewsVO> getNewsById(Long id);
 
     /**
      * 分页查询消息列表
